requirements.txt
streamlit
spacy
pdfkit
pillow
qrcode
requests
fuzzywuzzy